requirements.txt
asgiref==3.3.4
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
